Add automatic sorting of tasks in TaskList class
Tasks are automatically sorted as they are added to the task list when
the user enters.

Changes include: the updating of addTask() method to call Comparator to
sort the tasks.

This enhancement improves user experience and code quality as it does
not require the user to manually sort the task one by one.